Refactor experiment management components: - Moved NewExperimentDetailsCard, NewExperimentMembersCard, and NewExperimentVisibilityCard to a new directory structure. - Reimplemented NewExperimentDetailsCard, NewExperimentMembersCard, and NewExperimentVisibilityCard with updated logic and structure. - Created ExperimentDetailsCard, ExperimentInfoCard, and ExperimentVisibilityCard for managing existing experiments. - Added ExperimentSettings component to encapsulate experiment settings functionality. - Implemented hooks for experiment creation, updating, and deletion. - Enhanced user experience with improved form handling and validation using react-hook-form and zod.
