+++ apps/web/components/experiment-settings/index.ts
@@ -1,0 +1,1 @@
+export { ExperimentSettings } from "./experiment-settings";
