+++ apps/web/components/new-experiment/index.ts
@@ -1,0 +1,1 @@
+export { NewExperimentForm } from "./new-experiment";
